+++ src/CommandLine/GitLab/GitLabLocationLines.cs
@@ -1,0 +1,6 @@
+namespace Roslynator.CommandLine.GitLab;
+
+internal sealed class GitLabLocationLines
+{
+    public int Begin { get; set; }
+}